[add] adapterを分離
