+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/view/adapter/CustomAdapter.kt
@@ -1,0 +1,49 @@
+package jp.co.yumemi.android.code_check.view.adapter
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.TextView
+import androidx.recyclerview.widget.DiffUtil
+import androidx.recyclerview.widget.ListAdapter
+import androidx.recyclerview.widget.RecyclerView
+import jp.co.yumemi.android.code_check.R
+import jp.co.yumemi.android.code_check.domain.model.item.Item
+
+//差分を比較して計算する
+//変更された部分だけ更新できる
+val diff_util = object : DiffUtil.ItemCallback<Item>() {
+    override fun areItemsTheSame(oldItem: Item, newItem: Item): Boolean {
+        return oldItem.name == newItem.name
+    }
+
+    override fun areContentsTheSame(oldItem: Item, newItem: Item): Boolean {
+        return oldItem == newItem
+    }
+
+}
+
+class CustomAdapter(
+    private val itemClickListener: OnItemClickListener,
+) : ListAdapter<Item, CustomAdapter.ViewHolder>(diff_util) {
+
+    class ViewHolder(view: View) : RecyclerView.ViewHolder(view)
+
+    interface OnItemClickListener {
+        fun itemClick(item: Item)
+    }
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ViewHolder {
+        val view = LayoutInflater.from(parent.context).inflate(R.layout.layout_item, parent, false)
+        return ViewHolder(view)
+    }
+
+    override fun onBindViewHolder(holder: ViewHolder, position: Int) {
+        val item = getItem(position)
+        (holder.itemView.findViewById<View>(R.id.repositoryNameView) as TextView).text = item.name
+
+        holder.itemView.setOnClickListener {
+            itemClickListener.itemClick(item)
+        }
+    }
+}
